pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "dbt-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "salswl215",email = "lyn7666@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "dbt-glue (>=1.10.19,<2.0.0)",
    "dbt-spark[pyhive] (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
